+++ src/connectivity/mod.rs
@@ -24,12 +24,8 @@ pub mod controllable_object_message_data;
 pub mod dynamic_object_message_data;
 pub mod handlers;
 pub mod models;
-pub mod player_profile_handlers;
-pub mod player_profiles;
-pub mod player_session_handlers;
-pub mod player_sessions;
+pub mod player_info;
 pub mod server_client_message;
 pub mod services;
 pub mod user_session;
 pub mod view_layers;
-pub mod websocket_handler;
