+++ src/main.rs
@@ -38,10 +38,10 @@ use bevy_ecs::prelude::{Commands, Local, Res, Resource};
 use bevy_ecs::world::World;
 use clap::Parser;
 use connectivity::connected_users::ConnectingUsersQueue;
-use connectivity::player_profile_handlers::{create_new_ephemeral_player, create_new_username_player};
-use connectivity::player_profiles::PlayerProfiles;
-use connectivity::player_session_handlers::login_player;
-use connectivity::player_sessions::PlayerSessions;
+use connectivity::handlers::player_profile_handlers::{create_new_ephemeral_player, create_new_username_player};
+use connectivity::player_info::player_profiles::PlayerProfiles;
+use connectivity::handlers::player_session_handlers::login_player;
+use connectivity::player_info::player_sessions::PlayerSessions;
 use euclid::Angle;
 use futures::channel::mpsc::unbounded;
 use rand::Rng;
@@ -51,7 +51,7 @@ use tower_http::compression::CompressionLayer;
 use tracing::{debug, trace, Level};
 use tracing_subscriber::FmtSubscriber;
 
-use connectivity::websocket_handler::*;
+use connectivity::handlers::websocket_handler::*;
 use std::net::SocketAddr;
 use std::path::PathBuf;
 use std::sync::Arc;
